Change emptying policy
Empty a node even if aggregation makes the node smaller than what is
defined to be a full node. This is because, if we don't then the node
has to be compressed again and the parent has to check whether this
compression has completed before copying into the child. This causes
unnecessary complexity for no gain at all.