separation of PAO operations and data
The C-style separation is being done for memory-efficiency. This is
accomplished through:
* avoiding the vtable pointer overhead in each PAO
* constructing the PAO as key and value pointers. This allows only the value
  pointer to be stored in the hashtable while aggregating without maintaining
  information about the key, thus saving 8B of memory per hashtable entry.

We do this by separating the operations on PAOs into a singleton class (which
can be overloaded to specialize for applications) and the data into a PAO class
which consists of pointers to a string key and a Value-type value. Value can be
specialized again, but has no virtual functions and hence avoids overhead.

Contents:
* Update filters to support model
* Update aggregators
* Update the mapper and reducer
* Update sparsehash; still storing PAOs as values for now
* Leave out allowing tokens to include PAOs for now (don't quite remember why
  this was done)
* Clean up CBT linking information and don't install all headers