Simpler model
This model brings back PAOs. I do this to be able to use protobufs, which we'll
need for performance.  We still preserve the separation between operations and
PAO data, so the vtable pointer overhead is gone. In the case of hash-based
aggregation, we still point to the PAO, so I think what we can do to save
memory is to make the key pointer NULL in the PAO when it is entered in the
hashtable.  During eviction we can assign the key again from the hashtable.